Continued working on generate prelim targets code. Have a working version, but need to test with imported AI contours. Also started cleaning up code

+++ VMATTBICSIAutoPlanMT/VMATCSIAutoPlanMT/VMAT_CSI/planPrep_CSI.cs
@@ -47,9 +47,9 @@ namespace VMATCSIAutoPlanMT.VMAT_CSI
             int calcItems = 2;
             ProvideUIUpdate(0, "Initializing...");
             List<List<Beam>> beamsPerIso = PlanPrepHelper.ExtractBeamsPerIso(VMATPlan);
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved list of beams for each isocenter for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved list of beams for each isocenter for plan: {VMATPlan.Id}");
             List<string> isoNames = IsoNameHelper.GetCSIIsoNames(beamsPerIso.Count);
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved isocenter names for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved isocenter names for plan: {VMATPlan.Id}");
             ProvideUIUpdate($"Separating isocenters in plan {VMATPlan.Id} into separate plans");
             if (SeparateVMATPlan(VMATPlan, beamsPerIso, isoNames)) return true;
             ProvideUIUpdate(100, $"Successfully separated isocenters in plan {VMATPlan.Id}");
