+++ VMATTBICSIAutoPlanMT/VMATTBIAutoPlanMT/VMAT_TBI/PlanPrep_TBI.cs
@@ -67,7 +67,7 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             int calcItems = 2;
             ProvideUIUpdate(0, "Initializing...");
             List<List<Beam>> vmatBeamsPerIso = PlanPrepHelper.ExtractBeamsPerIso(VMATPlan);
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved list of beams for each isocenter for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved list of beams for each isocenter for plan: {VMATPlan.Id}");
 
             numVMATIsos = vmatBeamsPerIso.Count;
             List<List<Beam>> appaBeamsPerIso = new List<List<Beam>> { };
@@ -75,28 +75,28 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             {
                 appaBeamsPerIso = PlanPrepHelper.ExtractBeamsPerIso(appaPlan);
                 numIsos = appaBeamsPerIso.Count + numVMATIsos;
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved list of beams for each isocenter for plan: {appaPlan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved list of beams for each isocenter for plan: {appaPlan.Id}");
             }
 
             //get the isocenter names using the isoNameHelper class
             List<string> isoNames = new List<string>(IsoNameHelper.GetTBIVMATIsoNames(numVMATIsos, numIsos));
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved isocenter names for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved isocenter names for plan: {VMATPlan.Id}");
 
             if (appaPlan != null)
             {
                 isoNames.AddRange(IsoNameHelper.GetTBIAPPAIsoNames(numVMATIsos, numIsos));
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved isocenter names for plan: {appaPlan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved isocenter names for plan: {appaPlan.Id}");
             }
 
             ProvideUIUpdate($"Separating isocenters in plan {VMATPlan.Id} into separate plans");
             if (SeparateVMATPlan(VMATPlan, vmatBeamsPerIso, isoNames)) return true;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Successfully separated isocenters in plan {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Successfully separated isocenters in plan {VMATPlan.Id}");
 
             if (appaPlan != null)
             {
                 ProvideUIUpdate($"Separating isocenters in plan {appaPlan.Id} into separate plans");
                 if(SeparateAPPAPlan(appaPlan, appaBeamsPerIso, isoNames)) return true;
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Successfully separated isocenters in plan {appaPlan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Successfully separated isocenters in plan {appaPlan.Id}");
             }
             return false;
         }
@@ -113,11 +113,11 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
                 ExternalPlanSetup newplan = (ExternalPlanSetup)appaPlan.Course.CopyPlanSetup(appaPlan);
                 List<Beam> beamsToRemove = new List<Beam> { };
                 newplan.Id = String.Format("{0} {1}", count + 1, isoNames.ElementAt(count).Contains("upper") ? "Upper Legs" : "Lower Legs");
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Created new plan {newplan.Id} as copy of {appaPlan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Created new plan {newplan.Id} as copy of {appaPlan.Id}");
                 
                 //newplan.AddReferencePoint(newplan.StructureSet.Structures.First(x => x.Id.ToLower() == "ptv_body"), null, newplan.Id, newplan.Id);
                 separatedPlans.Add(newplan);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Added {newplan.Id} to list of separated plans");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Added {newplan.Id} to list of separated plans");
                 foreach (Beam b in newplan.Beams)
                 {
                     //if the current beam in newPlan is NOT found in the beams list, then remove it from the current new plan
@@ -127,10 +127,10 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
                         beamsToRemove.Add(b);
                     }
                 }
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished identifying beams that need to be removed from {newplan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished identifying beams that need to be removed from {newplan.Id}");
 
                 if (RemoveExtraBeams(newplan, beamsToRemove)) return true;
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Removed excess beams from {newplan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Removed excess beams from {newplan.Id}");
 
                 count++;
             }
@@ -145,17 +145,17 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             int percentComplete = 0;
             int calcItems = 6;
             List<ExternalPlanSetup> plans = new List<ExternalPlanSetup> { VMATPlan };
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Added VMAT plan ({VMATPlan.Id}) to list of plans requiring dose recalculation");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Added VMAT plan ({VMATPlan.Id}) to list of plans requiring dose recalculation");
 
             if (appaPlan != null)
             {
                 plans.Add(appaPlan);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Added APPA plan ({appaPlan.Id}) to list of plans requiring dose recalculation");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Added APPA plan ({appaPlan.Id}) to list of plans requiring dose recalculation");
             }
 
             (bool isError, List<ExternalPlanSetup> otherPlans) = CheckExistingPlansUsingSameSSWIthDoseCalculated(VMATPlan.Course.Patient.Courses.ToList(), VMATPlan.StructureSet);
             if (isError) return true;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished checking for existing plans that have dose calculated and use the same structure set!");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished checking for existing plans that have dose calculated and use the same structure set!");
 
             if (otherPlans.Any())
             {
@@ -164,13 +164,13 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
                 ProvideUIUpdate("Added found existing plans to list of plans requiring dose recalculation");
             }
             if (ResetCalculationMatrix(plans)) return true;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished resetting the dose calculation matrix for all applicable plans");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished resetting the dose calculation matrix for all applicable plans");
 
             if (RemoveFlashStructures()) return true;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished removing all structures used to create flash");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished removing all structures used to create flash");
 
             if (CopyHumanBodyOntoBody()) return true;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished copying human_body structure onto body");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished copying human_body structure onto body");
 
             return false;
         }
@@ -182,11 +182,11 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             ProvideUIUpdate("Removing flash structures");
 
             StructureSet ss = VMATPlan.StructureSet;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved structure set ({ss.Id}) used for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved structure set ({ss.Id}) used for plan: {VMATPlan.Id}");
 
             List<Structure> flashStr = ss.Structures.Where(x => x.Id.ToLower().Contains("flash") && !x.IsEmpty).ToList();
             calcItems += flashStr.Count;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved list of structures used to create flash");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved list of structures used to create flash");
             //List<Structure> removeMe = new List<Structure>(flashStr);
             //ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Copied list of flash structures to a dummy list that can be used for iteration");
 
@@ -195,10 +195,10 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             {
                 if (ss.CanRemoveStructure(itr))
                 {
-                    ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Removing structure {itr.Id} from structure set");
+                    ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Removing structure {itr.Id} from structure set");
                     ss.RemoveStructure(itr);
                 }
-                else ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Warning! Could not remove structure: {itr.Id}! Skipping!");
+                else ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Warning! Could not remove structure: {itr.Id}! Skipping!");
             }
             return false;
         }
@@ -209,17 +209,17 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
             int percentComplete = 0;
             int calcItems = 3;
             StructureSet ss = VMATPlan.StructureSet;
-            ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved structure set ({ss.Id}) used for plan: {VMATPlan.Id}");
+            ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved structure set ({ss.Id}) used for plan: {VMATPlan.Id}");
 
             //from the generateTS class, the human_body structure was a copy of the body structure BEFORE flash was added. Therefore, if this structure still exists, we can just copy it back onto the body
             if (StructureTuningHelper.DoesStructureExistInSS("human_body", ss, true))
             {
                 ProvideUIUpdate($"Human_body structure exists in structure set");
                 Structure body = StructureTuningHelper.GetStructureFromId("body", ss);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved body structure: {body.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved body structure: {body.Id}");
 
                 Structure bodyCopy = StructureTuningHelper.GetStructureFromId("human_body", ss);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Retrieved human_body structure: {bodyCopy.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Retrieved human_body structure: {bodyCopy.Id}");
 
                 (bool fail, StringBuilder message) = ContourHelper.CopyStructureOntoStructure(bodyCopy, body);
                 if(fail)
@@ -227,11 +227,11 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
                     ProvideUIUpdate(message.ToString(), true);
                     return true;
                 }
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Copied {bodyCopy.Id} onto {body.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Copied {bodyCopy.Id} onto {body.Id}");
 
                 if (ss.CanRemoveStructure(bodyCopy))
                 {
-                    ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Removing {bodyCopy.Id} structure from structure set");
+                    ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Removing {bodyCopy.Id} structure from structure set");
                     ss.RemoveStructure(bodyCopy);
                 }
                 else
@@ -284,7 +284,7 @@ namespace VMATTBIAutoPlanMT.VMAT_TBI
                 string calcModel = itr.GetCalculationModel(CalculationType.PhotonVolumeDose);
                 itr.ClearCalculationModel(CalculationType.PhotonVolumeDose);
                 itr.SetCalculationModel(CalculationType.PhotonVolumeDose, calcModel);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Reset dose calculation matrix for plan: {itr.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Reset dose calculation matrix for plan: {itr.Id}");
             }
             return false;
         }
