+++ VMATTBICSIAutoPlanMT/VMATTBICSIAutoplanningHelpers/BaseClasses/PlanPrepBase.cs
@@ -29,7 +29,7 @@ namespace VMATTBICSIAutoPlanningHelpers.BaseClasses
                     sb.AppendLine(b.Id);
                     beamFormatError = true;
                 }
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Beam {b.Id} formatting ok");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Beam {b.Id} formatting ok");
             }
             if (beamFormatError)
             {
@@ -64,13 +64,13 @@ namespace VMATTBICSIAutoPlanningHelpers.BaseClasses
                 //copy the plan, set the plan id based on the counter, and make a empty list to hold the beams that need to be removed
                 ExternalPlanSetup newplan = (ExternalPlanSetup)plan.Course.CopyPlanSetup(plan);
                 newplan.Id = $"{count + 1} {names.ElementAt(count)}";
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Created new plan {newplan.Id} as copy of {plan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Created new plan {newplan.Id} as copy of {plan.Id}");
                 List<Beam> beamsToRemove = new List<Beam> { };
                 //can't add reference point to plan because it must be open in Eclipse for ESAPI to perform this function. Need to fix in v16
                 //newplan.AddReferencePoint(newplan.StructureSet.Structures.First(x => x.Id.ToLower() == "ptv_body"), null, newplan.Id, newplan.Id);
                 //add the current plan copy to the separatedPlans list
                 separatedPlans.Add(newplan);
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Added {newplan.Id} to list of separated plans");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Added {newplan.Id} to list of separated plans");
                 //loop through each beam in the plan copy and compare it to the list of beams in the current isocenter
                 foreach (Beam b in newplan.Beams)
                 {
@@ -84,10 +84,10 @@ namespace VMATTBICSIAutoPlanningHelpers.BaseClasses
                         beamsToRemove.Add(b);
                     }
                 }
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Finished identifying beams that need to be removed from {newplan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Finished identifying beams that need to be removed from {newplan.Id}");
 
                 if (RemoveExtraBeams(newplan, beamsToRemove)) return true;
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Removed excess beams from {newplan.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Removed excess beams from {newplan.Id}");
 
                 count++;
             }
@@ -124,7 +124,7 @@ namespace VMATTBICSIAutoPlanningHelpers.BaseClasses
             {
                 ProvideUIUpdate($"Recalculating dose for plan: {p.Id}");
                 p.CalculateDose();
-                ProvideUIUpdate((int)(100 * ++percentComplete / calcItems), $"Dose recalculated for plan: {p.Id}");
+                ProvideUIUpdate(100 * ++percentComplete / calcItems, $"Dose recalculated for plan: {p.Id}");
             }
             return false;
         }
